requirements.txt
flask==3.0.0
streamlit==1.28.1
python-dotenv==1.0.0
groq==0.4.2
openai==1.3.0
cohere==4.47
google-generativeai==0.3.0
nltk==3.8.1
requests==2.31.0
gunicorn==21.2.0
